move asset generation logic to `Mix.Tasks.Compile.Surface.AssetGenerator`
